move icon and toggle in to livehelpers
