+++ lib/home_hub_web/live/live_helpers.ex
@@ -1,6 +1,6 @@
 defmodule HomeHubWeb.LiveHelpers do
   @moduledoc false
-  # import Phoenix.LiveView
+  import Phoenix.LiveView
   import Phoenix.LiveView.Helpers
 
   alias HomeHubWeb.Router.Helpers, as: Routes
@@ -40,6 +40,46 @@ defmodule HomeHubWeb.LiveHelpers do
         <% end %>
       </section>
     </header>
+    """
+  end
+
+  def toggle(assigns) do
+    assigns =
+      assigns
+      |> assign_new(:phx_click, fn -> "" end)
+      |> assign_new(:phx_target, fn -> nil end)
+
+    ~H"""
+    <div class="px-4 py-2" phx-click={@phx_click} phx-target={@phx_target}>
+      <%= if @on do %>
+        <img class="icon icon-on-orange" src={Routes.static_path(HomeHubWeb.Endpoint, "/images/icons/toggle-on.svg")}>
+      <% else %>
+        <img class="icon icon-action" src={Routes.static_path(HomeHubWeb.Endpoint, "/images/icons/toggle-off.svg")}>
+      <% end %>
+    </div>
+
+    <div class="px-4 py-2 text-4xl" phx-click={@phx_click} phx-target={@phx_target}>
+      <%= render_slot(@inner_block) %>
+    </div>
+    """
+  end
+
+  def icon(assigns) do
+    assigns =
+      assigns
+      |> assign_new(:on, fn -> false end)
+      |> assign_new(:on_colour, fn -> "" end)
+      |> assign_new(:off_colour, fn -> "icon-on-gray" end)
+      |> assign_new(:phx_click, fn -> nil end)
+      |> assign_new(:phx_target, fn -> nil end)
+
+    ~H"""
+    <div
+      class="px-4 py-2"
+      phx-click={@phx_click}
+      phx-target={@phx_target}>
+      <img class={"icon #{if assigns.on, do: assigns.on_colour, else: assigns.off_colour}"} src={@src} />
+    </div>
     """
   end
 end
